updatedcode

+++ core/models.py
@@ -67,7 +67,7 @@ class DomesticProduct(models.Model):
     productname = models.CharField(max_length=250)
     markup = models.IntegerField(verbose_name ='Mark Up Rate(in %)',default = 35,blank=True, null=True)
     productcostc = models.FloatField(default = 0.494,verbose_name = "Product Cost C$",null=True,blank=True)
-    targetgrossprofit = models.FloatField(default = 33,verbose_name = "Target Gross Profit (in %)",null=True,blank=True)
+    targetgrossprofit = models.IntegerField(default = 33,verbose_name = "Target Gross Profit (in %)",null=True,blank=True)
 
     def __str__(self):
         return self.sku
@@ -118,7 +118,7 @@ class ImportsProduct(models.Model):
     productname = models.CharField(max_length=250)
     setupfee = models.IntegerField(blank=True, null=True)
     markuprate = models.IntegerField(verbose_name ='Mark Up Rate(in %)',default = 35,blank=True, null=True)
-    targetgrossprofit = models.FloatField(verbose_name = "Target Gross Profit (in %)",null=True,blank=True)
+    targetgrossprofit = models.IntegerField(verbose_name = "Target Gross Profit (in %)",null=True,blank=True)
     duty = models.IntegerField(verbose_name ="Duty (in %)",default = 18,null=True,blank=True)
     markup = models.IntegerField(verbose_name ="Markup (in %)",default = 15,null=True,blank=True)
     frieghtvalue = models.IntegerField(verbose_name ="FRIEGHT ADMIN/UNIT (in %)",default = 15,null=True,blank=True)
